alteração dos exercícios
