+++ exercicios-obrigatorio/ExercicioS6-ToDoList_NaiaraLima.js
@@ -1,15 +1,14 @@
-
 const { request, response } = require('express');
 // Criar as APIs necessárias para as seguintes telas
 // Criar as chamadas das APis criadas no postman
 
 const bodyParse = require('body-parser') // Utilitário q serve para receber dados de qq formulário dentro do express 
-const express= require('express');
+const express= require('express'); // auxilia na construção das nossas aplicações Web
 
 
 const app = express();
 
-app.use(bodyParse.json())
+app.use(bodyParse.json()) // o use ele fica entre as chamadas >> bodyparse serve para fazer o express entender o json (no postman)
 
 
 const Port = 3001;
@@ -20,49 +19,6 @@ function listenFunction(){ // subindo o servidor >> o listen recebe dois paramet
 
 app.listen(Port,listenFunction)
 
-
-//tela de login
-
-function loginUser(request,response){
-    const login = [
-        {
-            name:'Naiara Lima',
-            email: 'naiarafx@gmail.com',
-            password: 'ontheway2020'
-        }
-    ]
-    return response.status(200).send(login) // dar uma resposta
-}
-
-// ** no postman colocar localhost:3001/login ** 
-app.get("/login",loginUser) 
-
-
-//-------------------------------------------------------------
-//tela de cadastro
-const newRegister = [{ // coloquei isso no postman
-    name: 'Tifany Lima',
-    email: 'tocadatify@gmail.com',
-    password: '666TapaNaPantera'
-  }]
-
-
-const UserRegister = (request,response) =>{
-    const newUser = request.body
-    console.log('USER', newUser)
-    newRegister.push(newUser)
-   if(newUser.name && newUser.email && newUser.password){
-        return response.status(201).send({messagem : 'Cadastro feito com sucesso'})
-    }else{
-        return response.status(400).send({ message: 'Erro no email ou senha inválida'})
-    }
-}
-
-// ** no postman colocar localhost:3001/login ** 
-app.post("/login",UserRegister) 
-
-
-//-------------------------------------------------------------
 // Tela de Listagem de TODO List
 
 const listagem = [{
@@ -80,7 +36,7 @@ const listagem = [{
 }]
 
 const toDoList = (request,response)=>response.status(200).send(listagem)
-app.get("/login",toDoList) 
+app.get("/listagem",toDoList) // mudar o  endereço de http pois são coisas diferentes
 
 
 //-------------------------------------------------------------
@@ -95,7 +51,7 @@ const newToDo = [{ // coloquei isso no postman
 
 const toDoRegister = (resquest,response)=>{
     const newToDoRegister = request.body
-    console.log('NewToDo', newToDoRegister)
+    console.log('NewToDo', newToDo)
     newToDo.push(newToDoRegister)
     if(newToDoRegister.title && newToDoRegister.text && newToDoRegister.id && newToDoRegister.status){
         return response.status(201).send({message: 'Cadastro feito com sucesso'})
@@ -104,10 +60,10 @@ const toDoRegister = (resquest,response)=>{
     }
 }
 
-app.post("/login", toDoRegister) 
+app.post("/listagem", toDoRegister) 
 
-//-------------------------------------------------------------
-// atualização de um ToDoList 
+// //-------------------------------------------------------------
+// // atualização de um ToDoList 
 
 const upDateToDo = (request,response)=>{
     const id = request.params.id // como boa prática é bom colocar o id do produto 
@@ -117,5 +73,4 @@ const upDateToDo = (request,response)=>{
         return response.status(400).send({message: 'Faltou enviar um dado'})
     }
 }
-app.put("/login", upDateToDo)
-
+app.put("/listagem/:id", upDateToDo)
